+++ rio_tiler/io/base.py
@@ -18,6 +18,7 @@ from ..errors import (
     TileOutsideBounds,
 )
 from ..expression import apply_expression
+from ..models import ImageStatistics, Info, Metadata, SpatialInfo
 from ..tasks import multi_arrays, multi_values
 
 
@@ -40,14 +41,14 @@ class SpatialMixin:
         )
 
     @property
-    def spatial_info(self) -> Dict:
+    def spatial_info(self) -> SpatialInfo:
         """Return Dataset's spatial info."""
-        return {
-            "bounds": self.bounds,
-            "center": self.center,
-            "minzoom": self.minzoom,
-            "maxzoom": self.maxzoom,
-        }
+        return SpatialInfo(
+            bounds=self.bounds,
+            center=self.center,
+            minzoom=self.minzoom,
+            maxzoom=self.maxzoom,
+        )
 
     def tile_exists(self, tile_z: int, tile_x: int, tile_y: int) -> bool:
         """Check if a tile is inside a the dataset bounds."""
@@ -74,21 +75,24 @@ class BaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractmethod
-    def info(self) -> Dict:
+    def info(self) -> Info:
         """Return Dataset's info."""
         ...
 
     @abc.abstractmethod
-    def stats(self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any) -> Dict:
+    def stats(
+        self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
+    ) -> Dict[str, ImageStatistics]:
         """Return Dataset's statistics."""
         ...
 
-    @abc.abstractmethod
-    def metadata(self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,) -> Dict:
+    def metadata(
+        self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
+    ) -> Metadata:
         """Return Dataset's statistics and info."""
         info = self.info()
-        info["statistics"] = self.stats(pmin, pmax, **kwargs)
-        return info
+        stats = self.stats(pmin, pmax, **kwargs)
+        return Metadata(statistics=stats, **info.dict())
 
     @abc.abstractmethod
     def tile(
@@ -128,26 +132,25 @@ class AsyncBaseReader(SpatialMixin, metaclass=abc.ABCMeta):
         pass
 
     @abc.abstractmethod
-    async def info(self) -> Coroutine[Any, Any, Dict]:
+    async def info(self) -> Coroutine[Any, Any, Info]:
         """Return Dataset's info."""
         ...
 
     @abc.abstractmethod
     async def stats(
         self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any
-    ) -> Coroutine[Any, Any, Dict]:
+    ) -> Coroutine[Any, Any, Dict[str, ImageStatistics]]:
         """Return Dataset's statistics."""
         ...
 
     async def metadata(
         self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any,
-    ) -> Coroutine[Any, Any, Dict]:
+    ) -> Coroutine[Any, Any, Metadata]:
         """Return Dataset's statistics and info."""
         info, stats = await asyncio.gather(
             *[self.info(), self.stats(pmin, pmax, **kwargs)]
         )
-        info["statistics"] = stats
-        return info
+        return Metadata(statistics=stats, **info.dict())
 
     @abc.abstractmethod
     async def tile(
@@ -199,9 +202,9 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
         _re = re.compile(assets.replace("\\\\", "\\"))
         return tuple(set(re.findall(_re, expression)))
 
-    def info(
+    def info(  # type: ignore
         self, assets: Union[Sequence[str], str] = None, *args, **kwargs: Any
-    ) -> Dict:
+    ) -> Dict[str, Info]:
         """Return metadata from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -216,13 +219,13 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
 
         return multi_values(assets, _reader, *args, **kwargs)
 
-    def stats(
+    def stats(  # type: ignore
         self,
         pmin: float = 2.0,
         pmax: float = 98.0,
         assets: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, Dict[str, ImageStatistics]]:
         """Return array statistics from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -237,13 +240,13 @@ class MultiBaseReader(BaseReader, metaclass=abc.ABCMeta):
 
         return multi_values(assets, _reader, pmin, pmax, **kwargs)
 
-    def metadata(
+    def metadata(  # type: ignore
         self,
         pmin: float = 2.0,
         pmax: float = 98.0,
         assets: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, Metadata]:
         """Return metadata from multiple assets"""
         if not assets:
             raise MissingAssets("Missing 'assets' option")
@@ -472,7 +475,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
 
     def info(
         self, bands: Union[Sequence[str], str] = None, *args, **kwargs: Any
-    ) -> Dict:
+    ) -> Info:
         """Return metadata from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -480,24 +483,28 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         if isinstance(bands, str):
             bands = (bands,)
 
-        def _reader(band: str, **kwargs: Any) -> Dict:
+        def _reader(band: str, **kwargs: Any) -> Info:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
                 return cog.info()
 
         bands_metadata = multi_values(bands, _reader, *args, **kwargs)
-        meta = self.spatial_info
+
+        meta = self.spatial_info.dict()
         meta["band_metadata"] = [
-            (ix + 1, bands_metadata[band]["band_metadata"][0][1])
+            (band, bands_metadata[band].band_metadata[0][1])
             for ix, band in enumerate(bands)
         ]
-        meta["band_descriptions"] = [(ix + 1, band) for ix, band in enumerate(bands)]
-        meta["dtype"] = bands_metadata[bands[0]]["dtype"]
-        meta["colorinterp"] = [
-            bands_metadata[band]["colorinterp"][0] for _, band in enumerate(bands)
+        meta["band_descriptions"] = [
+            (band, bands_metadata[band].band_descriptions[0][1])
+            for ix, band in enumerate(bands)
         ]
-        meta["nodata_type"] = bands_metadata[bands[0]]["nodata_type"]
-        return meta
+        meta["dtype"] = bands_metadata[bands[0]].dtype
+        meta["colorinterp"] = [
+            bands_metadata[band].colorinterp[0] for _, band in enumerate(bands)
+        ]
+        meta["nodata_type"] = bands_metadata[bands[0]].nodata_type
+        return Info(**meta)
 
     def stats(
         self,
@@ -505,7 +512,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         pmax: float = 98.0,
         bands: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Dict[str, ImageStatistics]:
         """Return array statistics from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -516,7 +523,9 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         def _reader(band: str, *args, **kwargs) -> Dict:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
-                return cog.stats(*args, **kwargs)[1]
+                return cog.stats(*args, **kwargs)[
+                    "band1"
+                ]  # We only return statistics for Band `1` of each dataset.
 
         return multi_values(bands, _reader, pmin, pmax, **kwargs)
 
@@ -526,7 +535,7 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         pmax: float = 98.0,
         bands: Union[Sequence[str], str] = None,
         **kwargs: Any,
-    ) -> Dict:
+    ) -> Metadata:
         """Return metadata from multiple bands"""
         if not bands:
             raise MissingBands("Missing 'bands' option")
@@ -534,30 +543,33 @@ class MultiBandReader(BaseReader, metaclass=abc.ABCMeta):
         if isinstance(bands, str):
             bands = (bands,)
 
-        def _reader(band: str, *args, **kwargs) -> Dict:
+        def _reader(band: str, *args, **kwargs) -> Metadata:
             url = self._get_band_url(band)
             with self.reader(url, tms=self.tms, **self.reader_options) as cog:  # type: ignore
-                meta = cog.metadata(*args, **kwargs)
-                meta["statistics"] = meta["statistics"][1]
-                return meta
+                return cog.metadata(*args, **kwargs)
 
         bands_metadata = multi_values(bands, _reader, pmin, pmax, **kwargs)
 
-        meta = self.spatial_info
+        meta = self.spatial_info.dict()
         meta["band_metadata"] = [
-            (ix + 1, bands_metadata[band]["band_metadata"][0][1])
+            (band, bands_metadata[band].band_metadata[0][1])
             for ix, band in enumerate(bands)
         ]
-        meta["band_descriptions"] = [(ix + 1, band) for ix, band in enumerate(bands)]
-        meta["dtype"] = bands_metadata[bands[0]]["dtype"]
-        meta["colorinterp"] = [
-            bands_metadata[band]["colorinterp"][0] for _, band in enumerate(bands)
+        meta["band_descriptions"] = [
+            (band, bands_metadata[band].band_descriptions[0][1])
+            for ix, band in enumerate(bands)
         ]
-        meta["nodata_type"] = bands_metadata[bands[0]]["nodata_type"]
+        meta["dtype"] = bands_metadata[bands[0]].dtype
+        meta["colorinterp"] = [
+            bands_metadata[band].colorinterp[0] for _, band in enumerate(bands)
+        ]
+        meta["nodata_type"] = bands_metadata[bands[0]].nodata_type
         meta["statistics"] = {
-            band: bands_metadata[band]["statistics"] for _, band in enumerate(bands)
+            # We only keep statistics for Band `1` of each dataset.
+            band: bands_metadata[band].statistics["band1"]
+            for _, band in enumerate(bands)
         }
-        return meta
+        return Metadata(**meta)
 
     def tile(
         self,