+++ tests/test_io_async.py
@@ -13,6 +13,7 @@ import pytest
 
 from rio_tiler import constants
 from rio_tiler.io import AsyncBaseReader, COGReader
+from rio_tiler.models import ImageStatistics, Info
 
 try:
     import contextvars  # Python 3.7+ only or via contextvars backport.
@@ -57,13 +58,13 @@ class AsyncCOGReader(AsyncBaseReader):
         self.minzoom = self.dataset.minzoom
         self.maxzoom = self.dataset.maxzoom
 
-    async def info(self) -> Coroutine[Any, Any, Dict]:
+    async def info(self) -> Coroutine[Any, Any, Info]:
         """Return Dataset's info."""
         return await run_in_threadpool(self.dataset.info)  # type: ignore
 
     async def stats(
         self, pmin: float = 2.0, pmax: float = 98.0, **kwargs: Any
-    ) -> Coroutine[Any, Any, Dict]:
+    ) -> Coroutine[Any, Any, Dict[str, ImageStatistics]]:
         """Return Dataset's statistics."""
         return await run_in_threadpool(self.dataset.stats, pmin, pmax, **kwargs)  # type: ignore
 
@@ -103,8 +104,8 @@ async def test_async():
         assert info == dataset.info()
 
         meta = cog.spatial_info
-        assert meta.get("minzoom") == 4
-        assert meta.get("maxzoom") == 8
+        assert meta.minzoom == 4
+        assert meta.maxzoom == 8
 
         assert await cog.stats(5, 95)
         assert await cog.metadata(2, 98)